GridBatch cleanup
Cleaned up the GridBatch python interface.

Grid and GridBatch are now fully immutable, all modifications create new objects.

Construction of a Grid or GridBatch now happens entirely through the GridBatch.from_* or Grid.from_* classmethods. These eliminate the need for an overload on __init__, and are much more clear in their intentions. 

More clearly defined the distinction between a GridBatch with batch_size (grid_count) of zero - which is "empty because of zero grids", vs a GridBatch (or Grid) in which individual grids exist but contain zero voxels. In the latter case, voxel_size and origin are still defined, as are the bounding boxes, whereas in the zero-grid-count case, they're undefined.

GridBatch has tightened the API so that it no longer auto-promotes arguments from Tensor to JaggedTensor. In the multi-grid case, where GridBatch should be used, this conversion shouldn't be implicit. For the single-grid case, just switch to Grid, and then you can use (and receive) tensors directly.

---------

Signed-off-by: Christopher Horvath <blackencino+git@gmail.com>