Point Encoder (#10)
Added a Pointnet-style Point Encoder which works on jagged tensor
inputs.

It has extensive unit tests against the synthetic street scenes we
create from the test utils.

It does not assume the existence of a SparseFeatureHierarchy, and
doesn't even expect a world coordinate frame. Instead, it assumes that
coordinate transformation is handled externally and that the points
passed in are in voxel coordinate space, such that floor() produces the
correct ijk coordinates.

---------

Signed-off-by: Christopher Horvath <chorvath@nvidia.com>